More V2 upgrade regression fixes, field alignment Web --> API --> DB

+++ Dtos/PlayerStatsDto.cs
@@ -18,7 +18,7 @@ public class PlayerStatsDto
     public decimal? AssistsPerGame { get; set; }
 
     // NFL
-    public int? Sacks { get; set; }
+    public decimal? Sacks { get; set; }
     public int? Touchdowns { get; set; }
 
     // NHL
